pyproject.toml
[tool.poetry]
name = "fast-tmp"
version = "0.1.0"
description = "为fastapi项目开发做准备"
authors = ["chise <chise123@live.com>"]
license = "Apache-2.0"
repository = "https://github.com/Chise1/fast-tmp"
packages = [
    { include = "fast_tmp" },
    { include = "cache" }
]
include = ["LICENSE", "README.md"]


[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.15.0"
fastapi = "^0.62.0"
tortoise-orm = "^0.16.18"
watchgod = "^0.6"
typer = "^0.3.2"
uvicorn = "^0.12.3"
python-multipart = "^0.0.5"
aerich = { git = "https://github.com/Chise1/aerich.git", branch = "dev" }


[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "*"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
dbdrivers = ["aioredis"]
sentry = ["sentry-sdk"]

[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true

#启动脚本

[tool.poetry.scripts]
fast-tmp = "fast_tmp.cli:main"
